insta_trend_suggester/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
apscheduler==3.10.4
instaloader==4.10.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
asyncpg==0.29.0
httpx==0.25.2
